Added new level counting system and backup before refactoring brick system for audio adding

+++ Assets/Scripts/PlayMode/Level/IReadOnlyLevelData.cs
@@ -5,6 +5,6 @@ namespace PlayMode.Level
 {
     public interface IReadOnlyLevelData
     {
-        public IReadOnlyReactiveProperty<int> Level { get; }
+        public IReadOnlyReactiveProperty<float> Level { get; }
     }
 }
